Refactor code for issue #560
