+++ Zilon.Core/Zilon.Core/Persons/BaseSurvivalData.cs
@@ -1,0 +1,123 @@
+﻿using System;
+using System.Linq;
+using Zilon.Core.Persons.Survival;
+
+namespace Zilon.Core.Persons
+{
+    public class BaseSurvivalData
+    {
+        public BaseSurvivalData(SurvivalStat[] stats)
+        {
+            Stats = stats;
+        }
+
+        /// <summary>Текущие характеристики.</summary>
+        public SurvivalStat[] Stats { get; }
+
+        /// <summary>
+        /// Событие, которое происходит, если значение характеристики
+        /// пересекает ключевое значение (мин/макс/четверти/0).
+        /// </summary>
+        public event EventHandler<SurvivalStatChangedEventArgs> StatChanged;
+
+        /// <summary>Признак того, что персонаж мёртв.</summary>
+        public bool IsDead { get; private set; }
+
+        /// <summary>Происходит, если персонаж умирает.</summary>
+        public event EventHandler Dead;
+
+        /// <summary>Пополнение запаса характеристики.</summary>
+        /// <param name="type">Тип характеритсики, которая будет произведено влияние.</param>
+        /// <param name="value">Значение, на которое восстанавливается текущий запас.</param>
+        public void RestoreStat(SurvivalStatType type, int value)
+        {
+            ValidateStatChangeValue(value);
+
+            var stat = Stats.SingleOrDefault(x => x.Type == type);
+            if (stat != null)
+            {
+                ChangeStatInner(stat, value);
+            }
+        }
+
+        /// <summary>Снижение характеристики.</summary>
+        /// <param name="type">Тип характеритсики, которая будет произведено влияние.</param>
+        /// <param name="value">Значение, на которое снижается текущий запас.</param>
+        public void DecreaseStat(SurvivalStatType type, int value)
+        {
+            ValidateStatChangeValue(value);
+
+            var stat = Stats.SingleOrDefault(x => x.Type == type);
+            if (stat != null)
+            {
+                ChangeStatInner(stat, -value);
+            }
+        }
+
+        /// <summary>
+        /// Invokes the stat changed event with the given parameters
+        /// </summary>
+        /// <param name="caller">The object performing the call</param>
+        /// <param name="args">The <see cref="SurvivalStatChangedEventArgs"/> instance containing the event data.</param>
+        public void InvokeStatChangedEvent(BaseSurvivalData caller, SurvivalStatChangedEventArgs args)
+        {
+            StatChanged?.Invoke(caller, args);
+        }
+
+        /// <summary>Форсированно установить запас здоровья.</summary>
+        /// <param name="type">Тип характеритсики, которая будет произведено влияние.</param>
+        /// <param name="value">Целевое значение запаса характеристики.</param>
+        public void SetStatForce(SurvivalStatType type, int value)
+        {
+            var stat = Stats.SingleOrDefault(x => x.Type == type);
+            if (stat != null)
+            {
+                stat.Value = value;
+            }
+        }
+
+        private void ValidateStatChangeValue(int value)
+        {
+            if (value == 0)
+            {
+                return;
+            }
+
+            if (value < 0)
+            {
+                throw new ArgumentException("Величина не может быть меньше 0.", nameof(value));
+            }
+        }
+
+        protected void ChangeStatInner(SurvivalStat stat, int value)
+        {
+            stat.Value += value;
+
+            ProcessIfHealth(stat);
+        }
+
+        /// <summary>
+        /// Индивидуально обрабатывает характеристику, если это здоровье.
+        /// </summary>
+        /// <param name="stat"> Обрабатываемая характеристика. </param>
+        private void ProcessIfHealth(SurvivalStat stat)
+        {
+            if (stat.Type != SurvivalStatType.Health)
+            {
+                return;
+            }
+
+            var hp = stat.Value;
+            if (hp <= 0)
+            {
+                IsDead = true;
+                DoDead();
+            }
+        }
+
+        private void DoDead()
+        {
+            Dead?.Invoke(this, new EventArgs());
+        }
+    }
+}
