+++ Zilon.Core/Zilon.Core/Persons/HumanSurvivalData.cs
@@ -12,20 +12,34 @@ namespace Zilon.Core.Persons
     /// <summary>
     /// Базовая реализация данных о выживании для человеческих персонажей.
     /// </summary>
-    public sealed class HumanSurvivalData : ISurvivalData
+    public sealed class HumanSurvivalData : BaseSurvivalData, ISurvivalData
     {
         private readonly IPersonScheme _personScheme;
         private readonly ISurvivalRandomSource _randomSource;
 
+//        /// <summary>
+//        /// Событие, которое происходит, если значение характеристики
+//        /// пересекает ключевое значение (мин/макс/четверти/0).
+//        /// </summary>
+//        public event EventHandler<SurvivalStatChangedEventArgs> StatChanged;
+
         public HumanSurvivalData([NotNull] IPersonScheme personScheme,
-            [NotNull] ISurvivalRandomSource randomSource)
+            [NotNull] ISurvivalRandomSource randomSource) : base(GetStats(personScheme))
         {
             _personScheme = personScheme ?? throw new ArgumentNullException(nameof(personScheme));
             _randomSource = randomSource ?? throw new ArgumentNullException(nameof(randomSource));
 
+            foreach (var stat in Stats)
+            {
+                stat.Changed += Stat_Changed;
+            }
+        }
+
+        private static SurvivalStat[] GetStats([NotNull] IPersonScheme personScheme)
+        {
             // Устанавливаем характеристики выживания персонажа
             var statList = new List<SurvivalStat>();
-            SetHitPointsStat(_personScheme, statList);
+            SetHitPointsStat(personScheme, statList);
 
             // Выставляем сытость/упоённость
             if (personScheme.SurvivalStats != null)
@@ -46,11 +60,7 @@ namespace Zilon.Core.Persons
                     statList);
             }
 
-            Stats = statList.ToArray();
-            foreach (var stat in Stats)
-            {
-                stat.Changed += Stat_Changed;
-            }
+            return statList.ToArray();
         }
 
         private void Stat_Changed(object sender, EventArgs e)
@@ -82,69 +92,13 @@ namespace Zilon.Core.Persons
 
         public HumanSurvivalData([NotNull] IPersonScheme personScheme,
             [NotNull] IEnumerable<SurvivalStat> stats,
-            [NotNull] ISurvivalRandomSource randomSource)
+            [NotNull] ISurvivalRandomSource randomSource) : base(stats.ToArray())
         {
             _personScheme = personScheme ?? throw new ArgumentNullException(nameof(personScheme));
             _randomSource = randomSource ?? throw new ArgumentNullException(nameof(randomSource));
 
-            Stats = stats.ToArray();
         }
-
-        /// <summary>Текущие характеристики.</summary>
-        public SurvivalStat[] Stats { get; }
-
-        /// <summary>Признак того, что персонаж мёртв.</summary>
-        public bool IsDead { get; private set; }
-
-        /// <summary>
-        /// Событие, которое происходит, если значение характеристики
-        /// пересекает ключевое значение (мин/макс/четверти/0).
-        /// </summary>
-        public event EventHandler<SurvivalStatChangedEventArgs> StatChanged;
-
-        /// <summary>Происходит, если персонаж умирает.</summary>
-        public event EventHandler Dead;
-
-        /// <summary>Пополнение запаса характеристики.</summary>
-        /// <param name="type">Тип характеритсики, которая будет произведено влияние.</param>
-        /// <param name="value">Значение, на которое восстанавливается текущий запас.</param>
-        public void RestoreStat(SurvivalStatType type, int value)
-        {
-            ValidateStatChangeValue(value);
-
-            var stat = Stats.SingleOrDefault(x => x.Type == type);
-            if (stat != null)
-            {
-                ChangeStatInner(stat, value);
-            }
-        }
-
-        /// <summary>Снижение характеристики.</summary>
-        /// <param name="type">Тип характеритсики, которая будет произведено влияние.</param>
-        /// <param name="value">Значение, на которое снижается текущий запас.</param>
-        public void DecreaseStat(SurvivalStatType type, int value)
-        {
-            ValidateStatChangeValue(value);
-
-            var stat = Stats.SingleOrDefault(x => x.Type == type);
-            if (stat != null)
-            {
-                ChangeStatInner(stat, -value);
-            }
-        }
-
-        /// <summary>Форсированно установить запас здоровья.</summary>
-        /// <param name="type">Тип характеритсики, которая будет произведено влияние.</param>
-        /// <param name="value">Целевое значение запаса характеристики.</param>
-        public void SetStatForce(SurvivalStatType type, int value)
-        {
-            var stat = Stats.SingleOrDefault(x => x.Type == type);
-            if (stat != null)
-            {
-                stat.Value = value;
-            }
-        }
-
+        
         /// <summary>Обновление состояния данных о выживании.</summary>
         public void Update()
         {
@@ -163,50 +117,6 @@ namespace Zilon.Core.Persons
                     ChangeStatInner(stat, -stat.Rate);
                 }
             }
-        }
-
-        private void ValidateStatChangeValue(int value)
-        {
-            if (value == 0)
-            {
-                return;
-            }
-
-            if (value < 0)
-            {
-                throw new ArgumentException("Величина не может быть меньше 0.", nameof(value));
-            }
-        }
-
-        private void ChangeStatInner(SurvivalStat stat, int value)
-        {
-            stat.Value += value;
-
-            ProcessIfHealth(stat);
-        }
-
-        /// <summary>
-        /// Индивидуально обрабатывает характеристику, если это здоровье.
-        /// </summary>
-        /// <param name="stat"> Обрабатываемая характеристика. </param>
-        private void ProcessIfHealth(SurvivalStat stat)
-        {
-            if (stat.Type != SurvivalStatType.Health)
-            {
-                return;
-            }
-
-            var hp = stat.Value;
-            if (hp <= 0)
-            {
-                IsDead = true;
-                DoDead();
-            }
-        }
-
-        private void DoDead()
-        {
-            Dead?.Invoke(this, new EventArgs());
         }
 
         private static SurvivalStat CreateStat(SurvivalStatType type, PersonSurvivalStatType schemeStatType, IPersonSurvivalStatSubScheme[] survivalStats)
@@ -261,9 +171,8 @@ namespace Zilon.Core.Persons
         private void DoStatChanged(SurvivalStat stat)
         {
             var args = new SurvivalStatChangedEventArgs(stat);
-            StatChanged?.Invoke(this, args);
+            InvokeStatChangedEvent(this, args);
         }
-
 
         private int GetStatDownRoll(SurvivalStat stat)
         {
